Very rudimentary "windowed" mode implemented.
This adds a dependency on the Minhook library. Due to its lack of a
proper build system, it is included as a git submodule with proper cmake
support linking it into our DLL as a static library. Due to the lack of
symbols for the IDirectDraw interface, I opted to hack each created
instance's vftable.

There are still a number of problems here... Namely:
- JMP3 expects to have an exclusive 16bpp drawing mode. There are
definitely some problems here, considering basically all systems run in
32bpp now.
- JMP3 has a tools menu that pops up when nearing the top of the screen.
When enabling a proper title bar, the title bar is hidden along with the
menu.
- JMP3 used an exclusive mode that painted unused pixels black. Because
we drop the exclusive mode request on the floor, those pixels are
presently transparent.