Implement proxy drawing surface for Legacy
Legacy expects to draw to a 16bpp primary surface. We now intercept this
request and provide a hidden surface for it to draw on. Unfortunately,
using DirectDraw to blit a 16bpp surface to a 32bpp is fairly slow.
Furthermore, any locking operation on a DDSurface acquires the process's
Win16 lock, causing major hitching type behavior. So, we have a thread
to convert 16bpp to 32bpp pixels and blit.

There are still a few problems:
- The draw thread has a few potential race conditions around quitting
- QuickTime videos are offset slightly toward the bottom. The QuickTime
DLL seems to be taking into account both HWND position and title bar
size
- QuickTime videos seem to play back in a halting manner at times
- The intro video does not render
- Moving the game window causes the cursor to behave oddly
- The cursor flickers