Fix #5
The inventory UI popup refuses to work if we have clipped the cursor
with `ClipCursor()`. So, we have to be a little bit smarter about our
handling of cursor capturing in that we only do so when the user is
dragging the cursor. There are interesting bugs that emerge when the
cursor leaves the game window while panning a node...