Implement frame timer and friends.
Now you can show the current FPS and frame time in game. This is mostly
for debugging purposes. While debugging and profiling, I determined that
forcing the primary ddraw surface to system memory results in a HUGE
performance boon. I expect this is because we are memcpy'ing it, after
all. This does seem to introduce some minor bugs with elements not being
refreshed until they are moused over. These are likely bugs in
directdraw or the driver itself, so I'm not sure what can be done. They
are very minor and happen very rarely, so...