Fix drawing issues introduced by 588a7c0.
I'm still uncertain if this is a bug in DirectDraw that happens when a
surface is allocated in system memory or if it is merely a side effect
of our faster draw loop. Either way, it appears that DirectDraw will not
handle multiple threads requesting a surface lock in this case.
DirectDraw will return DDERR_SURFACEBUSY. The solution is to gate
access to the proxy surface behind an std::recursive_mutex, which allows
the main thread to acquire multiple locks on the surface -- remember
that IDirectDrawSurface::Lock can lock multiple sections of the surface.

The bugs are fixed and performance impact appears minimal.